checks for deploys

+++ bin/redeploy.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
 
-# kill old process
-id=`pgrep -f run.rb`
-cmd="kill -9 $id"
-eval $cmd
+LOCAL=$(git rev-parse @)
+REMOTE=$(git rev-parse @{u})
+BASE=$(git merge-base @ @{u})
 
-# pull from repo
-git pull origin master
-bundle install
-rake db:migrate
+if [ $LOCAL = $REMOTE ]; then
+    echo "Up-to-date"
+elif [ $LOCAL = $BASE ]; then
+  # kill old process
+  id=`pgrep -f run.rb`
+  cmd="kill -9 $id"
+  eval $cmd
 
-# run in background
-echo `nohup ./run.rb 0<&- &> logs/log.txt &`
+  # pull from repo
+  git pull origin master
+  bundle install
+  rake db:migrate
+
+  # run in background
+  echo `nohup ./run.rb 0<&- &> logs/log.txt &`
+fi
